+++ src/extension.ts
@@ -3,90 +3,12 @@
 import * as vscode from "vscode"
 import * as https from "https"
 
-const getCommentOpening = (editor: vscode.TextEditor | undefined) => {
-	switch (editor?.document.languageId) {
-		case "c":
-		case "cpp":
-		case "csharp":
-		case "cuda-cpp":
-		case "objective-c":
-		case "objective-cpp":
-		case "go":
-		case "rust":
-		case "java":
-		case "scala":
-		case "kotlin":
-		case "scss":
-		case "swift":
-		case "typescript":
-		case "javascript":
-		case "javascriptreact":
-		case "coffeescript":
-		case "groovy":
-		case "php":
-			return "/*"
-		case "python":
-			return '"""'
-		case "ruby":
-			return "=begin"
-		case "perl":
-		case "perl6":
-			return "=pod"
-		case "r":
-			return "#"
-		default:
-			return "/*"
-	}
-}
-const getCommentEnding = (editor: vscode.TextEditor | undefined) => {
-	switch (editor?.document.languageId) {
-		case "c":
-		case "cpp":
-		case "csharp":
-		case "cuda-cpp":
-		case "objective-c":
-		case "objective-cpp":
-		case "go":
-		case "rust":
-		case "java":
-		case "scala":
-		case "kotlin":
-		case "scss":
-		case "swift":
-		case "typescript":
-		case "javascript":
-		case "javascriptreact":
-		case "coffeescript":
-		case "groovy":
-		case "php":
-			return /\*\/$/m
-		case "python":
-			return /""""$/m
-		case "ruby":
-			return /=end$/m
-		case "perl":
-		case "perl6":
-			return /=cut$/m
-		case "r":
-			return /#$/m
-		default:
-			return /\*\/$/m
-	}
-}
-
-const sendQueryToOpenAI = async (
-	queryText: string,
-	openAIKey: string,
-	commentClosing?: RegExp,
-) => {
+const sendQueryToOpenAI = async (queryText: string, openAIKey: string) => {
 	// get user or workspace configuration object
 	let config = vscode.workspace.getConfiguration()
-	const model = config.get("chatgpt-helper.model") as string | null
-	const modelId = model === "chatgpt" ? "gpt-3.5-turbo" : "code-davinci-002"
-	const modelNameReadable = model === "chatgpt" ? "ChatGPT" : "Codex"
 
 	const outputDocument = await vscode.workspace.openTextDocument({
-		content: `Querying ${modelNameReadable}. This may take some time...`,
+		content: `Querying ChatGPT. This may take some time...`,
 		language: "markdown",
 	})
 	const outputDocumentEditor = await vscode.window.showTextDocument(
@@ -104,8 +26,7 @@ const sendQueryToOpenAI = async (
 	const request = https.request(
 		{
 			hostname: "api.openai.com",
-			path:
-				model === "codex" ? "/v1/completions" : "/v1/chat/completions",
+			path: "/v1/chat/completions",
 			method: "POST",
 			headers: {
 				"Content-Type": "application/json", // eslint-disable-line @typescript-eslint/naming-convention
@@ -115,48 +36,29 @@ const sendQueryToOpenAI = async (
 		response => {
 			response.on("data", data => {
 				process.stdout.write(data)
-				if (model === "codex") {
-					responseText += JSON.parse(
-						data.toString(),
-					).choices[0].text.trim()
-					if (commentClosing) {
-						responseText = responseText.replace(commentClosing, "")
-					}
-				} else {
-					responseText += JSON.parse(data.toString()).choices[0]
-						.message.content
-				}
+				responseText += JSON.parse(data.toString()).choices[0].message
+					.content
 			})
 		},
 	)
 
 	request.write(
-		JSON.stringify(
-			model === "codex"
-				? {
-						model: "code-davinci-002",
-						prompt: entireQueryText,
-						temperature: 0.1,
-						max_tokens: 512, // eslint-disable-line @typescript-eslint/naming-convention
-						frequency_penalty: 0.38, // eslint-disable-line @typescript-eslint/naming-convention
-				  }
-				: {
-						model: "gpt-3.5-turbo",
-						max_tokens: 512, // eslint-disable-line @typescript-eslint/naming-convention
-						// todo: should subtract length (in tokens) of query
-						messages: [
-							{
-								role: "system",
-								content:
-									"You are a knowledgeable and helpful assistant called ChatGPT, who is specialised to help answer queries about users' code.",
-							},
-							{
-								role: "user",
-								content: entireQueryText,
-							},
-						],
-				  },
-		),
+		JSON.stringify({
+			model: "gpt-3.5-turbo",
+			max_tokens: 512, // eslint-disable-line @typescript-eslint/naming-convention
+			// todo: should subtract length (in tokens) of query
+			messages: [
+				{
+					role: "system",
+					content:
+						"You are a knowledgeable and helpful assistant called ChatGPT, who is specialised to help answer queries about users' code.",
+				},
+				{
+					role: "user",
+					content: entireQueryText,
+				},
+			],
+		}),
 	)
 	request.end()
 	request.on("close", () => {
@@ -179,7 +81,7 @@ const sendQueryToOpenAI = async (
 					new vscode.Position(0, 0),
 					new vscode.Position(99999999999999, 0),
 				),
-				`Error querying ${modelNameReadable}.\n\n${error}`,
+				`Error querying ChatGPT.\n\n${error}`,
 			)
 		})
 	})
@@ -225,32 +127,9 @@ export function activate(context: vscode.ExtensionContext) {
 			const codeToQuery = selectedCode || entireFileContents
 
 			if (codeToQuery) {
-				const workspaceConfiguration =
-					vscode.workspace.getConfiguration()
-				const commentOpening = getCommentOpening(
-					vscode.window.activeTextEditor,
-				)
-				const commentEnding = getCommentEnding(
-					vscode.window.activeTextEditor,
-				)
-				const model = workspaceConfiguration.get(
-					"chatgpt-helper.model",
-				) as string
 				const fullQueryText =
-					model === "chatgpt"
-						? "Why is the following code not working?\n" +
-						  codeToQuery
-						: codeToQuery +
-						  "\n" +
-						  commentOpening +
-						  (workspaceConfiguration.get(
-								"chatgpt-helper.codex.queries.codeNotWorking",
-						  ) as string | null)
-				sendQueryToOpenAI(
-					fullQueryText,
-					await getOpenAIKey(),
-					commentEnding,
-				)
+					"Why is the following code not working?\n" + codeToQuery
+				sendQueryToOpenAI(fullQueryText, await getOpenAIKey())
 			} else {
 				vscode.window.showErrorMessage(
 					"No code selected or file is empty. Did not send to ChatGPT",
@@ -277,31 +156,9 @@ export function activate(context: vscode.ExtensionContext) {
 			const codeToQuery = selectedCode || entireFileContents
 
 			if (codeToQuery) {
-				const workspaceConfiguration =
-					vscode.workspace.getConfiguration()
-				const commentOpening = getCommentOpening(
-					vscode.window.activeTextEditor,
-				)
-				const commentEnding = getCommentEnding(
-					vscode.window.activeTextEditor,
-				)
-				const model = workspaceConfiguration.get(
-					"chatgpt-helper.model",
-				) as string
 				const fullQueryText =
-					model === "chatgpt"
-						? "Explain the following code:\n" + codeToQuery
-						: codeToQuery +
-						  "\n" +
-						  commentOpening +
-						  (workspaceConfiguration.get(
-								"chatgpt-helper.codex.queries.explainCode",
-						  ) as string | null)
-				sendQueryToOpenAI(
-					fullQueryText,
-					await getOpenAIKey(),
-					commentEnding,
-				)
+					"Explain the following code:\n" + codeToQuery
+				sendQueryToOpenAI(fullQueryText, await getOpenAIKey())
 			} else {
 				vscode.window.showErrorMessage(
 					"No code selected or file is empty. Did not send to ChatGPT",